+++ packages/component/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @orderly.network/components
 
+## 0.0.7
+
+### Patch Changes
+
+- fix bug & update
+- Updated dependencies
+  - @orderly.network/types@0.0.7
+  - @orderly.network/net@1.0.5
+  - @orderly.network/utils@0.0.2
+
 ## 0.0.6
 
 ### Patch Changes