+++ packages/core/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @orderly.network/core
 
+## 0.0.7
+
+### Patch Changes
+
+- fix bug & update
+- Updated dependencies
+  - @orderly.network/types@0.0.7
+
 ## 0.0.6
 
 ### Patch Changes