+++ packages/hooks/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @orderly.network/hooks
 
+## 0.0.8
+
+### Patch Changes
+
+- fix bug & update
+- Updated dependencies
+  - @orderly.network/futures@0.0.8
+  - @orderly.network/types@0.0.7
+  - @orderly.network/core@0.0.7
+  - @orderly.network/net@1.0.5
+  - @orderly.network/utils@0.0.2
+
 ## 0.0.7
 
 ### Patch Changes