+++ packages/types/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @orderly.network/types
 
+## 0.0.7
+
+### Patch Changes
+
+- fix bug & update
+
 ## 0.0.6
 
 ### Patch Changes